% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioacousticsquote.R
\name{bioacousticsquote}
\alias{bioacousticsquote}
\alias{print.bioacousticsquote}
\alias{as.data.frame.bioacousticsquote}
\title{Display a randomly chosen statistical quote.}
\usage{
bioacousticsquote(
  ind = NULL,
  pattern = NULL,
  tag = NULL,
  source = NULL,
  topic = NULL
)

\method{print}{bioacousticsquote}(x, cite = TRUE, width = NULL, ...)

\method{as.data.frame}{bioacousticsquote}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{ind}{Integer or character.
If 'ind' is missing, a random quote is chosen from all quotations.
If 'ind' is specified and is an integer, return the ind^th quote.
If 'ind' is specified and is character, use it as the 'pattern'.}

\item{pattern}{Character string. Quotes are subset to to those which
match the pattern in the quote text.}

\item{tag}{Character string. Quotes are subset to those matching the
specified tag.}

\item{source}{Character string. Quotes are subset to those matching
the specified source (person).}

\item{topic}{Deprecated. Use 'tag' instead. Only kept for backward compatibility.}

\item{x}{object of class \code{'bioacousticsquote'}}

\item{cite}{logical; should the \code{cite} field be printed?}

\item{width}{Optional print width parameter}

\item{...}{Other optional arguments, unused here}

\item{row.names}{see \code{\link{as.data.frame}}}

\item{optional}{see \code{\link{as.data.frame}}}
}
\value{
A character vector containing one quote.
It is of class \code{bioacousticsquote} for which an S3 print method will be invoked, and for which
other methods are available.
}
\description{
Display a randomly chosen statistical quote.
}
\examples{
set.seed(1234)
bioacousticsquote()
bioacousticsquote(10)
bioacousticsquote("Seuketat")
bioacousticsquote(pattern="Seuketat")
bioacousticsquote(source="Hempton")
bioacousticsquote(tag="poetry")
print.data.frame(bioacousticsquote(9)) # All information

}
\seealso{
\code{\link{quote_tags}}, \code{\link{search_quotes}}, \code{\link{quotes}},
Inspired by: \code{\link[fortunes]{fortune}}

\code{\link{as.latex}}, \code{\link{as.markdown}}
}
